Fix missing hostvars.localhost.ansible_machine_id
When touching a new touchstone, it's contents are set based on details
obtained from the control_host (executing Ansible).  In some situations
the inventory name of this host may not be 'localhost'.  When this
occurs, the play will fail.

Fix this by adding a variable, defaulting to 'localhost', and ensure
it's facts are gathered prior to reference.  Update the README with
details about this variable and it's purpose.

Signed-off-by: Chris Evich <cevich@redhat.com>