Fix complexity + stderr/stdout merging
The Ansible script module merges the script's stderr/stdout before
relaying it back to the control-host.  Previously tasks erroneously
relied on this to determine 'changed' state and to set the (critical)
``stone_touched`` fact.  This was causing both to sometimes be
incorrect.

Fix this in the touchstone script, tasks, and reduce overall complexity
by utilizing a proper jinja2 template for the touchstone file contents.

Signed-off-by: Chris Evich <cevich@redhat.com>